requirements.txt
ccxt>=4.1.13
pyyaml>=6.0.1
rich>=13.7.0
pytest>=7.4.3
python-dotenv>=1.0.0
